backend/requirements.txt
Flask>=3.0
flask-cors>=4.0
Werkzeug>=3.0

pydub>=0.25.1
PyAudio==0.2.14; platform_system == "Windows"

numpy>=1.26
faster-whisper>=1.0.3
pyannote.audio>=3.1
pyannote.metrics>=3.2
torch>=2.2

groq>=0.9
python-docx>=1.1
reportlab>=4.1
jiwer>=3.0
rouge-score>=0.1.2